refactor(elementor-v4): split initialize.js into focused modules
Extracts document-tree.js, editor-data.js, change-handler.js, and
marks.js from initialize.js, leaving it as a thin entry point that
only wires Elementor hooks.
